pyproject.toml
[tool.poetry]
name = "llamatry"
version = "0.1.4"
description = "opentelemetry instrumentation for openai's completions api"
authors = ["Jason <jason@jxnl.co>"]
readme = "README.md"
repository = "https://github.com/jxnl/llamatry"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.6"
opentelemetry-api = "^1.17.0"
opentelemetry-sdk = "^1.17.0"
opentelemetry-instrumentation = "^0.38b0"
urllib3 = "<2.0"
opentelemetry-instrumentation-requests = "^0.38b0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
